pyproject.toml
[build-system]
# We use setuptools (the default, widely-supported backend).
# `build` is the PEP 517 frontend: `pip install build && python -m build`
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name        = "cicd-audit-framework"
version     = "1.0.0"
description = "Automated security and efficiency auditor for CI/CD pipelines"
readme      = "README.md"
requires-python = ">=3.11"
license     = { text = "MIT" }

keywords = [
  "ci", "cd", "cicd", "devops", "security", "audit",
  "github-actions", "gitlab-ci", "pipeline",
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Software Development :: Build Tools",
]

# Runtime dependencies — kept intentionally minimal
dependencies = [
  "PyYAML>=6.0",         # YAML parsing
  "Jinja2>=3.1",         # HTML report templating
  "click>=8.1",          # CLI framework
]

[project.optional-dependencies]
# Install with: pip install -e ".[dev]"
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4",           # fast linter + formatter
  "mypy>=1.10",          # static type checking
]

[project.scripts]
# After `pip install -e .` this command is available in your PATH
cicd-audit = "cicd_auditor.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["cicd_auditor*"]

[tool.setuptools.package-data]
# Ensure the Jinja2 template is included in the installed package
"cicd_auditor.reporter" = ["templates/*.j2"]

# ── Ruff (linter + formatter) ─────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length    = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter

# ── Mypy (type checking) ──────────────────────────────────────────────────
[tool.mypy]
python_version         = "3.11"
strict                 = false
ignore_missing_imports = true

# ── Pytest ────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths    = ["tests"]
addopts      = "-v --tb=short"
